change types, so swig can handle them

+++ software/libraries/C/openkubus.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-int32_t openkubus_authenticate(const char *pad, const char *pw, uint16_t offset, uint16_t num);
-void openkubus_gen_pad(const char *pw, uint16_t offset, uint16_t num, char *pad);
+int openkubus_authenticate(const char *pad, const char *pw, int offset, int num);
+void openkubus_gen_pad(const char *pw, int offset, int num, char *pad);
 
 #endif //OPENKUBUS_H
